requirements.txt
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
pytz==2022.1
sqlparse==0.4.4
